+++ x-pack/plugins/lens/public/datasources/form_based/operations/definitions/rate.tsx
@@ -21,8 +21,9 @@ import {
   getSafeName,
   getFilter,
   combineErrorMessages,
+  isMetricCounterField,
+  getErrorForRateReference,
 } from './helpers';
-import { isMetricCounterField } from '../../utils';
 
 const supportedTypes = new Set(['number']);
 
@@ -70,8 +71,7 @@ export const rateOperation: OperationDefinition<RateIndexPatternColumn, 'field',
   getErrorMessage: (layer, columnId, indexPattern) =>
     combineErrorMessages([
       getInvalidFieldMessage(layer, columnId, indexPattern),
-      //   getDisallowedPreviousShiftMessage(layer, columnId),
-      //   getColumnReducedTimeRangeError(layer, columnId, indexPattern),
+      getErrorForRateReference(layer, columnId, 'rate', indexPattern),
     ]),
   isTransferable: (column, newIndexPattern) => {
     const newField = newIndexPattern.getFieldByName(column.sourceField);
